+++ tests/conftest.py
@@ -6,9 +6,9 @@ from sqlalchemy.pool import StaticPool
 
 from app.auth import get_password_hash
 from app.database import Base, get_db
-from app.main import app
-from app.models import  User
 from app.enums import Region, Site, UserRole
+from app.main import app
+from app.models import User
 
 SQLALCHEMY_DATABASE_URL = "sqlite:///:memory:"